SPARK-740 Driver metrics (#25)
## What changes were proposed in this pull request?

Implements metrics in the coarse-grained driver implementation. The deprecated fine-grained driver is left as-is. Metrics include e.g. frequency of RPCs to/from Mesos, the internal state tracking of tasks/agents, and durations for bringing up the tasks.

## How was this patch tested?

Checked it by hand in a few deployments, and the numbers appear to all add up with nothing missing.